small fixes

+++ cadv_exploration/error_injection/corrupts/categorical_value_missing.py
@@ -4,7 +4,7 @@ from cadv_exploration.error_injection.basis import TabularCorruption
 
 
 class MissingCategoricalValueCorruption(TabularCorruption):
-    def __init__(self, columns: object = None, severity: int = 0.1, corrupt_strategy: object = "to_nan",
+    def __init__(self, columns: object = None, severity: float = 0.1, corrupt_strategy: object = "to_nan",
                  **kwargs: object):
         super().__init__(columns, **kwargs)
         self.severity = severity
